+++ pp/routing/add_electrical_pads_top_dc.py
@@ -48,6 +48,6 @@ if __name__ == "__main__":
     import pp
 
     c = pp.components.mzi2x2(with_elec_connections=True)
-    c = pp.components.wg_heater_connected()
+    c = pp.components.straight_with_heater()
     cc = add_electrical_pads_top_dc(component=c)
     cc.show()