+++ pp/routing/route_pad_array.py
@@ -170,7 +170,7 @@ def route_pad_array(
 
 if __name__ == "__main__":
 
-    c = pp.components.wg_heater_connected()
+    c = pp.components.straight_with_heater()
     c = pp.components.mzi2x2(with_elec_connections=True)
 
     elements, pads, _ = route_pad_array(c, fanout_length=100)