+++ CHANGELOG.md
@@ -33,8 +33,9 @@ from pydantic import validate_arguments
 - add Pdk.add_pins()
 - replace pp.write_gds(component, gdspath) with component.write_gds(gdspath)
 - replace pp.write_component(component, gdspath) with component.write_gds_with_metadata(gdspath)
-- rename pp.components.waveguide with
+- rename pp.components.waveguide to pp.components.straight
 - rename auto_taper_to_wide_waveguides auto_widen
+- rename wg_heater_connected to straight_with_heater
 
 
 ## 2.4.6